requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-allauth>=0.54.0
django-cors-headers>=4.0.0
django-filter>=23.0
django-extensions>=3.2.0
Pillow>=10.0.0
celery>=5.3.0
redis>=4.5.0
requests>=2.31.0
python-decouple>=3.8
django-crispy-forms>=2.0
crispy-tailwind>=0.5.0
djangorestframework-simplejwt>=5.2.0
django-guardian>=2.4.0
geopy>=2.3.0
scrapy>=2.9.0
selenium>=4.10.0
flower>=2.0.0
django-storages>=1.13.0
boto3>=1.26.0
sentry-sdk>=1.25.0